requirements.txt
django
gunicorn
django-heroku
requests
djangorestframework
stripe>=2.48.0,<3.0